folder restructuring and export fn fix causing build fails
