+++ app/admin/dashboard/profile/page.tsx
@@ -7,10 +7,10 @@ import {
   CardHeader,
   CardTitle,
   CardDescription,
-} from "@/app/dashboard/ui/card";
+} from "@/components/ui/card";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { AdminHeader } from "@/components/ui/adminheader";
-import { Button } from "@/app/dashboard/ui/button";
+import { Button } from "@/components/ui/button";
 import {
   UserCircle,
   Mail,
