+++ app/student/dashboard/notifications.tsx
@@ -1,4 +1,4 @@
-import { Card, CardContent, CardHeader, CardTitle } from "./ui/card"
+import { Card, CardContent, CardHeader, CardTitle } from "../../../components/ui/card"
 import { Bell } from "lucide-react"
 
 const notifications = [
