+++ app/student/dashboard/sidebar.tsx
@@ -11,7 +11,7 @@ import {
   GraduationCap,
 } from "lucide-react";
 import Link from "next/link";
-import { Button } from "./ui/button";
+import { Button } from "../../../components/ui/button";
 import { useState, useEffect } from "react";
 import { cn } from "@/lib/utils";
 import { logoutUser } from "@/lib/logout";
