+++ app/teacher/dashboard/notices/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import { Header } from "@/components/ui/teacherheader"
-import { Card, CardContent, CardHeader, CardTitle, CardDescription } from "@/app/dashboard/ui/card";
+import { Card, CardContent, CardHeader, CardTitle, CardDescription } from "@/components/ui/card";
 import Link  from "next/link";
 interface Notification {
     id: number;
