+++ components/ui/adminsidebar.tsx
@@ -1,3 +1,4 @@
+"use client";
 import {
   Home,
   Megaphone,
