+++ components/ui/teachersidebar.tsx
@@ -1,3 +1,4 @@
+"use client";
 import {
   Home,
   Calendar,
