+++ components/ui/use-mobile.tsx
@@ -1,3 +1,4 @@
+"use client";
 import * as React from "react"
 
 const MOBILE_BREAKPOINT = 768
